app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic>=2.10.1
openai>=1.52.0
requests>=2.32.0
